Count 'z' as five characters during decoding
When encoding if there is a chunk of all zeroes instead of being encoded
as "`!!!!!`" it is encoded as "`z`" as a lightweight form of
compression. This means that when decoding any "`z`" in the input needs
to be count as five characters (instead of one) when determining what
the length of the final output will be.

Since this was being counted wrong there was an overflow error when `z`
was being decoded because the buffer was smaller than it needed to be.

Fixes #1